requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pika==1.1.0
psycopg2-binary==2.8.4
Werkzeug==0.16.0
